+Small Gaps are filled as intended. Working perfectly now. I did need to change the Room vPosition data type from int to float. This did not affect our previous code besides creating a few conversion warnings to take care of.

+++ DungeonGenerator/Dungeon.h
@@ -22,7 +22,7 @@ public:
 
 	const int getNumCells()const { return nCells; };
 	const std::vector<Room*>& getRooms()const { return vRooms; };
-	const std::map<std::pair<int, int>, bool>& getTileMap()const { return tileMap; };
+	const std::map < std::pair<float, float > , bool > & getTileMap()const { return tileMap; };
 
 	void generateCellCoordinates();
 	void generateCellRectangles();
@@ -49,7 +49,7 @@ private:
 	int nDungeonSize;//Length/Width(in 'tiles') of the Dungeon. Area will be a square for simplicity.
 
 	/*tileMap used to indicate which 'tiles' of the Dungeon contain a tile. Use bool or int. Int may be useful to indicate the type of tile. 0 for empty. */
-	std::map<std::pair<int, int>, bool> tileMap;	//maps if a point contains part of a cell of some kind.
+	std::map < std::pair < float, float >, bool > tileMap;	//maps if a point contains part of a cell of some kind.
 	std::vector<Room*> vRooms;	//keeps track of the rooms.
 };
